+++ wxc_sdk/__init__.py
@@ -20,7 +20,7 @@ from dataclasses import dataclass
 
 __all__ = ['WebexSimpleApi']
 
-__version__ = '1.3.0'
+__version__ = '1.4.0'
 
 log = logging.getLogger(__name__)